move hooks.ts from components to rating module root for better organization
